requirements.txt
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
langchain-nvidia-ai-endpoints==0.0.11
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.0
requests==2.31.0
python-dotenv==1.0.0
